refactor(agent-task): keep lifecycle store path generic
