+++ src/core/lifecycle_store.rs
